docker_requirements.txt
# FastAPI app dependencies
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==1.10.14

# ML and data
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.7.0
joblib==1.4.2

# MLflow model loading
mlflow==2.13.0
cloudpickle==3.0.0

# AWS SDK
boto3==1.34.109

# Dotenv support
python-dotenv==1.0.1

# Model dependencies
cffi==1.17.1
lightgbm==4.6.0
psutil==7.0.0
scipy==1.16.0
